2.0 library - organize_data_generic - do not fail on missing files

+++ play.it-2/src/organize_data.sh
@@ -29,8 +29,10 @@ organize_data_generic() {
 	(
 		cd "$archive_path"
 		for file in $archive_files; do
-			mkdir --parents "$pkg_path/${file%/*}"
-			mv "$file" "$pkg_path/$file"
+			if [ -e "$file" ]; then
+				mkdir --parents "$pkg_path/${file%/*}"
+				mv "$file" "$pkg_path/$file"
+			fi
 		done
 	)
 }
